+++ tests/test_upload_rtc.py
@@ -1,3 +1,5 @@
+import json
+from collections import Counter
 from pathlib import Path
 from zipfile import ZipFile
 
@@ -6,13 +8,12 @@ from hyp3lib import aws
 from moto import mock_aws
 from moto.core import patch_client
 
-from hyp3_opera_rtc.upload_rtc import make_zip_name, upload_rtc
+from hyp3_opera_rtc import upload_rtc
 
 
-def test_upload_rtc(rtc_results_dir, rtc_output_files, s3_bucket):
+def test_upload_burst_rtc(rtc_burst_results_dir, s3_bucket):
     prefix = 'myPrefix'
-
-    upload_rtc(s3_bucket, prefix, rtc_results_dir)
+    upload_rtc.upload_rtc(s3_bucket, prefix, rtc_burst_results_dir)
 
     resp = aws.S3_CLIENT.list_objects_v2(Bucket=s3_bucket, Prefix=prefix)
 
@@ -24,53 +25,75 @@ def test_upload_rtc(rtc_results_dir, rtc_output_files, s3_bucket):
 
     assert zip_filename == f'{product_name}.zip'
 
-    zip_download_path = rtc_results_dir / 'output.zip'
+    zip_download_path = rtc_burst_results_dir / 'output.zip'
     aws.S3_CLIENT.download_file(s3_bucket, zip_s3_key, zip_download_path)
 
     with ZipFile(zip_download_path) as zf:
         files_in_zip = set([f.filename for f in zf.infolist()])
 
-        assert files_in_zip == set(
-            [
-                f'{product_name}/',
-                f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_BROWSE.png',
-                f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0.iso.xml',
-                f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0.h5',
-                f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_mask.tif',
-                f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_VH.tif',
-                f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_VV.tif',
-            ]
-        )
+        assert files_in_zip == {
+            f'{product_name}/',
+            f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_BROWSE.png',
+            f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0.iso.xml',
+            f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0.h5',
+            f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_mask.tif',
+            f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_VH.tif',
+            f'{product_name}/OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_VV.tif',
+        }
 
 
-def test_make_zip_name(rtc_output_files):
-    zip_filename = make_zip_name([Path(f) for f in rtc_output_files])
+def test_upload_slc_rtc(rtc_slc_results_dir, s3_bucket):
+    prefix = 'myPrefix'
+    upload_rtc.upload_rtc(s3_bucket, prefix, rtc_slc_results_dir)
+
+    resp = aws.S3_CLIENT.list_objects_v2(Bucket=s3_bucket, Prefix=prefix)
+
+    zip_s3_keys = [c['Key'] for c in resp['Contents'] if c['Key'].endswith('.zip')]
+    assert len(zip_s3_keys) == 0
+
+    uploaded_files = [c['Key'] for c in resp['Contents']]
+
+    assert len(uploaded_files) == 164
+    file_suffixs = dict(Counter(Path(f).suffix for f in uploaded_files))
+    assert file_suffixs == {'.json': 1, '.log': 1, '.h5': 27, '.xml': 27, '.png': 27, '.tif': 27 * 3}
+
+
+def test_make_zip_name(rtc_burst_output_files):
+    zip_filename = upload_rtc.make_zip_name([Path(f) for f in rtc_burst_output_files])
 
     assert zip_filename == 'OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0'
 
 
 @pytest.fixture
-def rtc_results_dir(tmp_path, rtc_output_files):
+def rtc_burst_results_dir(tmp_path, rtc_burst_output_files):
     (tmp_path / 'burst-dir').mkdir(parents=True)
 
-    for file in rtc_output_files:
+    for file in rtc_burst_output_files:
         (tmp_path / file).touch()
 
     return tmp_path
 
 
 @pytest.fixture
-def rtc_output_files():
-    return [
-        'OPERA_L2_RTC-S1_20250411T185446Z_S1A_30_v1.0.catalog.json',
-        'OPERA_L2_RTC-S1_20250411T185446Z_S1A_30_v1.0.log',
-        'OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0.iso.xml',
-        'OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_BROWSE.png',
-        'OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0.h5',
-        'OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_mask.tif',
-        'OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_VH.tif',
-        'OPERA_L2_RTC-S1_T115-245714-IW1_20240809T141633Z_20250411T185446Z_S1A_30_v1.0_VV.tif',
-    ]
+def rtc_slc_results_dir(tmp_path, rtc_slc_output_files):
+    (tmp_path / 'burst-dir').mkdir(parents=True)
+
+    for file in rtc_slc_output_files:
+        (tmp_path / file).touch()
+
+    return tmp_path
+
+
+@pytest.fixture
+def rtc_burst_output_files():
+    with (Path(__file__).parent / 'data' / 'rtc_output_files.json').open() as f:
+        return json.load(f)['burst']
+
+
+@pytest.fixture
+def rtc_slc_output_files():
+    with (Path(__file__).parent / 'data' / 'rtc_output_files.json').open() as f:
+        return json.load(f)['slc']
 
 
 @pytest.fixture